Adventure work products sales and budget analysis
Adventure Work Products Sales and Budget Analysis :-
This project involves the analysis of sales data for Adventure Work Products using Microsoft Power BI. The goal is to create a dashboard that provides insights into the sales performance of different product categories and subcategories across various countries, including the UK, US, Germany, Australia, France, and Canada.

Dashboard Overview :-
The dashboard presents a comprehensive view of the sales analysis, allowing users to understand the performance of the main categories: accessories, bikes, and clothing. Additionally, it provides insights into the sales of their respective subcategories.
Data Source :-
The data used in this analysis covers the period from January 2016 to December 2016. Ensure that your sales data is formatted appropriately and matches the required structure for the dashboard to function correctly.